Epic journey pt. 10
This is a big update. I have updated the interinsics and notation in chapters 1 and 2 completely. Heisenberg functions are more intuitive, using the tensor category when necessary. Documentation is updated up to the end of chapter 2. I inlcuded a python script that will search the intrinsics of the magma files and compare them to the doc files, located in tests/.