Induce changes to subspace closures
With the change to how Induce works and the change to the output of DerivationAlgebra when there exists a nontrivial partition, we fixed many bugs in DerivationClosure and NucleusClosure. I believe the intrinsics work like they did before. One big change is for the DerivationClosure and similar intrinsics (this is reflected in the documentation). If a 3-tensor t has coordinates that are fused, the input for Delta in DerivationClosure can come from the direct product of all 3 endo-spaces or a representative subset. Maybe we will include an example of this eventually.