Updates for Nuclei and Centroids
Updated the Nucleus algorithm. It is now much simpler and also slightly faster. The (worst-case) complexity of the algorithm has not changed. This change also updates the centroid algorithm and the cooresponding 3-tensor optimizations including Left-, Mid-, and RightNucleus. A new intrinsic has been added (no documentation yet) called `Centroid'. Now it can take either one argument or two: a tensor or a tensor with a set of coordinates A. It outputs the centroid and the A-centroid, respectively. More in documentation later.