Updates to DerivationAlgebra
Updated the DerivationAlgebra solver in a similar way to the Nucleus and Centroid solver. This removes the specialized algorithms for 3-tensors as the current algorithm works better for any valence. Two new intrinsics have been added: 'DerivationAlgebra' and 'SelfAdjointAlgebra'. The former allows for an additional argument---a set of coordinates. The latter is a function that computes the subspace of the nucleus which are fixed by the involution. More details will be in the documentation soon. Centroids also, by default, take the fusion rules of the category into account; the nucleus never does.