Added major functions
Added Shuffle for Hmtp. It should work just like Shuffle for tensors. Added a small fix to Shuffle on tensors: if the permutation is trivial, it just returns the input. AsCotensorSpace keeps the basis equivalent to SystemOfForms or the rows of the 0-foliation. Finally, you can call Valence on Hmtp.