Added more information in error statement
Every time a black-box tensor is created, we test the functions at 0 to make sure they can be evaluated. Previously if there was an error, it would just report that the functions could not be evaluated. Now they show the error report, which might help the user fix their function.